Remove dead code and retired migration paths
- Delete SessionManager.requestNotificationPermission, the never-produced
  PanelAction.navigatePanel, openInEditor(project:), and the unread
  notificationSessionPIDKey; copyPath reuses NSPasteboard.copyToClipboard.
- Delete SessionData.withSessionId and move the six snapshot-only qa
  fixture arrays into the test target.
- Retire the v0.12 single-position panel migration and the pre-0.6 no-PID
  cleanup branch in cctop-hook; records without a PID are left to the
  app's legacy-file sweep.
- Retire the pre-release manual-hide stable-key migration: manual hides
  are keyed by cctop_session_id only, and the visibility freeze now
  applies solely to partial inventories. Tests that exercised the
  migration are removed and docs/session-files.md describes the
  remaining contract.
