+++ menubar/CctopMenubar/Views/PopupViewHelpers.swift
@@ -60,8 +60,7 @@ extension PopupView {
     }
 
     func copyPath(_ path: String) {
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(path, forType: .string)
+        NSPasteboard.copyToClipboard(path)
     }
 }
