+++ menubar/CctopMenubarTests/SessionManagerVisibilityTests.swift
@@ -290,543 +290,6 @@ final class SessionManagerVisibilityTests: XCTestCase {
     }
 
     @MainActor
-    func testLegacyCodexManualHideMigratesBeforePublishingAndSurvivesReloads() throws {
-        let root = NSTemporaryDirectory() + "cctop-legacy-manual-hide-\(UUID().uuidString)"
-        let sessionsDir = (root as NSString).appendingPathComponent("sessions")
-        let historyDir = (root as NSString).appendingPathComponent("history")
-        try FileManager.default.createDirectory(atPath: sessionsDir, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(atPath: historyDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(atPath: root) }
-
-        let identityBlockerPath = (root as NSString).appendingPathComponent("session-identities")
-        try Data("not a directory".utf8).write(to: URL(fileURLWithPath: identityBlockerPath))
-        let threadID = "aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee"
-        let currentRecordID = "current-record"
-        let conflictingRecordID = "conflicting-record"
-        let cctopSessionID = "11111111-1111-4111-8111-111111111111"
-        let conflictingCctopSessionID = "22222222-2222-4222-8222-222222222222"
-        var session = SessionData(
-            sessionId: threadID,
-            projectPath: "/tmp/legacy-hidden",
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        session.cctopSessionId = nil
-        session.harnessSessionId = threadID
-        session.source = SessionData.codexSource
-        session.pid = 4_204
-        session.status = .waitingInput
-        let sessionPath = (sessionsDir as NSString).appendingPathComponent("codex-\(threadID).json")
-        try session.writeToFile(path: sessionPath)
-        var current = SessionData(
-            sessionId: currentRecordID,
-            projectPath: "/tmp/legacy-hidden",
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        current.cctopSessionId = cctopSessionID
-        current.harnessSessionId = threadID
-        current.source = SessionData.codexSource
-        current.pid = 4_205
-        current.status = .waitingInput
-        try current.writeToFile(
-            path: (sessionsDir as NSString).appendingPathComponent("codex-\(currentRecordID).json")
-        )
-        var conflicting = current
-        conflicting.sessionId = conflictingRecordID
-        conflicting.cctopSessionId = conflictingCctopSessionID
-        let conflictingPath = (sessionsDir as NSString).appendingPathComponent("codex-\(conflictingRecordID).json")
-        try conflicting.writeToFile(path: conflictingPath)
-
-        let suiteName = "cctop-legacy-manual-hide-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let visibility = ManualSessionVisibilityStore(defaults: defaults)
-
-        var sources = isolatedSessionDataSources(sessionsDir: URL(fileURLWithPath: sessionsDir), manualSessionVisibility: visibility)
-        sources.codexThreads = StubCodexThreadState(
-            existing: [threadID, currentRecordID, conflictingRecordID],
-            archived: []
-        )
-        sources.processAlive = { _ in true }
-        var postedNotificationIDs: [String] = []
-        sources.notificationsEnabled = { true }
-        sources.notificationClient = SessionNotificationClient(
-            add: { request, completion in
-                postedNotificationIDs.append(request.identifier)
-                completion(nil)
-            },
-            removePending: { _ in },
-            removeDelivered: { _ in }
-        )
-        let historyManager = HistoryManager(historyDir: URL(fileURLWithPath: historyDir))
-        let manager = SessionManager(historyManager: historyManager, dataSources: sources, startMonitoring: false)
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(postedNotificationIDs.isEmpty)
-        XCTAssertTrue(visibility.hiddenSessionIDs.isEmpty)
-        XCTAssertNil(try SessionData.fromFile(path: sessionPath).cctopSessionId)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-
-        try FileManager.default.removeItem(atPath: conflictingPath)
-        manager.loadSessions()
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(postedNotificationIDs.isEmpty)
-        XCTAssertEqual(visibility.hiddenSessionIDs, [cctopSessionID])
-        XCTAssertNil(try SessionData.fromFile(path: sessionPath).cctopSessionId)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-
-        try FileManager.default.removeItem(atPath: identityBlockerPath)
-        manager.loadSessions()
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertEqual(visibility.hiddenSessionIDs, [cctopSessionID])
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        XCTAssertTrue(postedNotificationIDs.isEmpty)
-        waitForIdentityMigrationQueue(manager)
-
-        manager.loadSessions()
-        XCTAssertNil(defaults.object(forKey: ManualSessionVisibilityStore.legacyDefaultsKey))
-        manager.loadSessions()
-        XCTAssertTrue(manager.userSessions.isEmpty)
-
-        let reloaded = SessionManager(historyManager: historyManager, dataSources: sources, startMonitoring: false)
-        XCTAssertTrue(reloaded.userSessions.isEmpty)
-        XCTAssertEqual(visibility.hiddenSessionIDs, [cctopSessionID])
-        XCTAssertTrue(postedNotificationIDs.isEmpty)
-    }
-
-    @MainActor
-    func testUnresolvedArchivedCodexLegacyHideKeepsCleanupAvailable() throws {
-        let root = NSTemporaryDirectory() + "cctop-unresolved-legacy-archived-\(UUID().uuidString)"
-        let sessionsDir = (root as NSString).appendingPathComponent("sessions")
-        let historyDir = (root as NSString).appendingPathComponent("history")
-        try FileManager.default.createDirectory(atPath: sessionsDir, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(atPath: historyDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(atPath: root) }
-
-        let identityBlockerPath = (root as NSString).appendingPathComponent("session-identities")
-        try Data("not a directory".utf8).write(to: URL(fileURLWithPath: identityBlockerPath))
-        let threadID = "bbbbbbbb-cccc-4ddd-8eee-ffffffffffff"
-        let projectPath = URL(fileURLWithPath: #filePath)
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .path
-        var session = codexSession(sessionId: threadID, projectPath: projectPath)
-        session.cctopSessionId = nil
-        session.harnessSessionId = threadID
-        session.endedAt = Date(timeIntervalSince1970: 2_100)
-        let sessionPath = (sessionsDir as NSString).appendingPathComponent("codex-\(threadID).json")
-        try session.writeToFile(path: sessionPath)
-        let unrelatedThreadID = "cccccccc-dddd-4eee-8fff-000000000001"
-        let unrelatedProjectPath = (root as NSString).appendingPathComponent("unrelated-project")
-        try FileManager.default.createDirectory(atPath: unrelatedProjectPath, withIntermediateDirectories: true)
-        var previous = SessionData(
-            sessionId: "previous-hidden-project",
-            projectPath: projectPath,
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        previous.endedAt = Date(timeIntervalSince1970: 2_000)
-        try previous.writeToFile(path: (historyDir as NSString).appendingPathComponent("previous-hidden-project.json"))
-        let unrelated = codexSession(
-            sessionId: unrelatedThreadID,
-            projectPath: unrelatedProjectPath
-        )
-        try unrelated.writeToFile(
-            path: (sessionsDir as NSString).appendingPathComponent("codex-\(unrelatedThreadID).json")
-        )
-
-        let suiteName = "cctop-unresolved-legacy-archived-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let visibility = ManualSessionVisibilityStore(defaults: defaults)
-
-        var sources = isolatedSessionDataSources(sessionsDir: URL(fileURLWithPath: sessionsDir), manualSessionVisibility: visibility)
-        sources.codexThreads = StubCodexThreadState(
-            existing: [threadID, unrelatedThreadID],
-            archived: [threadID, unrelatedThreadID]
-        )
-        sources.desktopAppConnection = DesktopAppConnectionLookup { _ in true }
-        sources.processAlive = { _ in true }
-        let manager = SessionManager(
-            historyManager: HistoryManager(historyDir: URL(fileURLWithPath: historyDir)),
-            dataSources: sources,
-            startMonitoring: false
-        )
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(
-            manager.cleanupSources.filter { $0.sessionId == threadID }.map(\.projectPath),
-            [projectPath]
-        )
-        XCTAssertEqual(
-            manager.cleanupSources.filter { $0.sessionId == unrelatedThreadID }.map(\.projectPath),
-            [unrelatedProjectPath]
-        )
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertTrue(visibility.hiddenSessionIDs.isEmpty)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-
-        let brokenPath = (sessionsDir as NSString).appendingPathComponent("broken.json")
-        try FileManager.default.removeItem(atPath: identityBlockerPath)
-        try Data("not valid session json".utf8).write(to: URL(fileURLWithPath: brokenPath))
-        manager.loadSessions()
-
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 1)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == unrelatedThreadID }.count, 1)
-        XCTAssertEqual(visibility.hiddenSessionIDs.count, 1)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        waitForIdentityMigrationQueue(manager)
-
-        try FileManager.default.removeItem(atPath: brokenPath)
-        manager.loadSessions()
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertNil(defaults.object(forKey: ManualSessionVisibilityStore.legacyDefaultsKey))
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 1)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == unrelatedThreadID }.count, 1)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-    }
-
-    @MainActor
-    func testProcessScopedLegacyKeyDoesNotFreezeArchivedProjectionsDuringPartialInventory() throws {
-        let root = NSTemporaryDirectory() + "cctop-process-key-partial-\(UUID().uuidString)"
-        let sessionsDir = (root as NSString).appendingPathComponent("sessions")
-        let historyDir = (root as NSString).appendingPathComponent("history")
-        try FileManager.default.createDirectory(atPath: sessionsDir, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(atPath: historyDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(atPath: root) }
-
-        let threadID = "dddddddd-eeee-4fff-8000-111111111111"
-        let projectPath = (root as NSString).appendingPathComponent("project")
-        try codexSession(sessionId: threadID, projectPath: projectPath)
-            .writeToFile(path: (sessionsDir as NSString).appendingPathComponent("codex-\(threadID).json"))
-        try Data("not valid session json".utf8).write(
-            to: URL(fileURLWithPath: (sessionsDir as NSString).appendingPathComponent("broken.json"))
-        )
-
-        let suiteName = "cctop-process-key-partial-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["active:42"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let visibility = ManualSessionVisibilityStore(defaults: defaults)
-
-        var sources = isolatedSessionDataSources(sessionsDir: URL(fileURLWithPath: sessionsDir), manualSessionVisibility: visibility)
-        sources.codexThreads = StubCodexThreadState(existing: [threadID], archived: [threadID])
-        sources.desktopAppConnection = DesktopAppConnectionLookup { _ in true }
-        sources.processAlive = { _ in true }
-        let manager = SessionManager(
-            historyManager: HistoryManager(historyDir: URL(fileURLWithPath: historyDir)),
-            dataSources: sources,
-            startMonitoring: false
-        )
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.map(\.projectPath), [projectPath])
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["active:42"]
-        )
-    }
-
-    @MainActor
-    func testUnresolvedLegacyFinishedSessionKeepsEvidenceAndRecentFrozen() throws {
-        let root = NSTemporaryDirectory() + "cctop-unresolved-legacy-finished-\(UUID().uuidString)"
-        let sessionsDir = (root as NSString).appendingPathComponent("sessions")
-        let historyDir = (root as NSString).appendingPathComponent("history")
-        try FileManager.default.createDirectory(atPath: sessionsDir, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(atPath: historyDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(atPath: root) }
-
-        let identityBlockerPath = (root as NSString).appendingPathComponent("session-identities")
-        try Data("not a directory".utf8).write(to: URL(fileURLWithPath: identityBlockerPath))
-        let projectPath = URL(fileURLWithPath: #filePath)
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .path
-        var history = SessionData(
-            sessionId: "previous-finished-project",
-            projectPath: projectPath,
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        history.endedAt = Date(timeIntervalSince1970: 2_000)
-        try history.writeToFile(path: (historyDir as NSString).appendingPathComponent("history.json"))
-
-        let threadID = "11111111-2222-4333-8444-555555555555"
-        var finished = SessionData(
-            sessionId: threadID,
-            projectPath: projectPath,
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        finished.cctopSessionId = nil
-        finished.harnessSessionId = threadID
-        finished.source = SessionData.codexSource
-        finished.pid = 4_205
-        finished.status = .working
-        let sessionPath = (sessionsDir as NSString).appendingPathComponent("codex-\(threadID).json")
-        try finished.writeToFile(path: sessionPath)
-        let fastPeerID = "finished-peer-fast"
-        var fastPeer = finished
-        fastPeer.sessionId = fastPeerID
-        fastPeer.cctopSessionId = "00000000-0000-4000-8000-000000000001"
-        fastPeer.pid = 4_206
-        let fastPeerPath = (sessionsDir as NSString).appendingPathComponent("codex-\(fastPeerID).json")
-        try fastPeer.writeToFile(path: fastPeerPath)
-
-        let codexPeerID = "finished-peer-desktop"
-        let codexPeerProjectPath = (root as NSString).appendingPathComponent("desktop-peer-project")
-        try FileManager.default.createDirectory(atPath: codexPeerProjectPath, withIntermediateDirectories: true)
-        var codexPeer = codexSession(sessionId: codexPeerID, projectPath: codexPeerProjectPath)
-        codexPeer.cctopSessionId = "00000000-0000-4000-8000-000000000002"
-        codexPeer.harnessSessionId = threadID
-        codexPeer.lastActivity = Date(timeIntervalSince1970: 2_100)
-        codexPeer.disconnectedAt = Date(timeIntervalSince1970: 2_100)
-        codexPeer.endedAt = Date(timeIntervalSince1970: 2_100)
-        let codexPeerPath = (sessionsDir as NSString).appendingPathComponent("codex-\(codexPeerID).json")
-        try codexPeer.writeToFile(path: codexPeerPath)
-
-        let legacyPeerID = "finished-peer-legacy"
-        var legacyPeer = codexSession(sessionId: legacyPeerID, projectPath: projectPath)
-        legacyPeer.cctopSessionId = "00000000-0000-4000-8000-000000000003"
-        legacyPeer.harnessSessionId = threadID
-        legacyPeer.lastActivity = Date(timeIntervalSince1970: 2_100)
-        legacyPeer.disconnectedAt = Date(timeIntervalSince1970: 2_100)
-        legacyPeer.endedAt = Date(timeIntervalSince1970: 2_100)
-        let legacyPeerPath = (sessionsDir as NSString).appendingPathComponent(
-            "33333333-3333-4333-8333-333333333333.json"
-        )
-        try legacyPeer.writeToFile(path: legacyPeerPath)
-        let finishedPeerIDs = [fastPeerID, codexPeerID, legacyPeerID]
-        let finishedPeerPaths = [fastPeerPath, codexPeerPath, legacyPeerPath]
-
-        let suiteName = "cctop-unresolved-legacy-finished-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let visibility = ManualSessionVisibilityStore(defaults: defaults)
-
-        var sources = isolatedSessionDataSources(sessionsDir: URL(fileURLWithPath: sessionsDir), manualSessionVisibility: visibility)
-        sources.codexThreads = StubCodexThreadState(existing: Set([threadID] + finishedPeerIDs), archived: [])
-        sources.desktopAppConnection = DesktopAppConnectionLookup { _ in false }
-        sources.processAlive = { _ in false }
-        let manager = SessionManager(
-            historyManager: HistoryManager(historyDir: URL(fileURLWithPath: historyDir)),
-            dataSources: sources,
-            startMonitoring: false
-        )
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertTrue(manager.cleanupSources.contains { $0.projectPath == projectPath })
-        XCTAssertEqual(
-            manager.cleanupSources.filter { $0.sessionId == codexPeerID }.map(\.projectPath),
-            [codexPeerProjectPath]
-        )
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertTrue(finishedPeerPaths.allSatisfy(FileManager.default.fileExists(atPath:)))
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-        XCTAssertTrue(visibility.hiddenSessionIDs.isEmpty)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-
-        manager.garbageCollectFinished()
-        XCTAssertTrue(finishedPeerPaths.allSatisfy(FileManager.default.fileExists(atPath:)))
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-        try finishedPeerPaths.forEach { try FileManager.default.removeItem(atPath: $0) }
-
-        manager.loadSessions()
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 0)
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-
-        let readyPath = (root as NSString).appendingPathComponent("identity-lock-ready")
-        let lockHolder = try startSessionLockHolder(
-            lockPath: sessionPath + ".lock",
-            readyPath: readyPath,
-            holdSeconds: 10
-        )
-        defer { terminateProcess(lockHolder) }
-
-        try FileManager.default.removeItem(atPath: identityBlockerPath)
-        manager.loadSessions()
-        XCTAssertNil(manager.sessionFileCache[sessionPath]?.session.cctopSessionId)
-        waitForIdentityMigrationQueue(manager)
-
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        let migratedSessionIDs = visibility.hiddenSessionIDs
-        XCTAssertEqual(migratedSessionIDs.count, 1)
-        XCTAssertNil(try SessionData.fromFile(path: sessionPath).cctopSessionId)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 0)
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-
-        manager.loadSessions()
-        waitForIdentityMigrationQueue(manager)
-
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertEqual(visibility.hiddenSessionIDs, migratedSessionIDs)
-        XCTAssertNil(try SessionData.fromFile(path: sessionPath).cctopSessionId)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 0)
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-
-        terminateProcess(lockHolder)
-        manager.loadSessions()
-        waitForIdentityMigrationQueue(manager)
-        XCTAssertEqual(try SessionData.fromFile(path: sessionPath).cctopSessionId, migratedSessionIDs.first)
-
-        manager.loadSessions()
-
-        XCTAssertNil(defaults.object(forKey: ManualSessionVisibilityStore.legacyDefaultsKey))
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertEqual(visibility.hiddenSessionIDs, migratedSessionIDs)
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(manager.cleanupSources.filter { $0.sessionId == threadID }.count, 0)
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-
-        defaults.removeObject(forKey: ManualSessionVisibilityStore.defaultsKey)
-        defaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        try FileManager.default.removeItem(atPath: sessionPath)
-        try fastPeer.writeToFile(path: fastPeerPath)
-        try codexPeer.writeToFile(path: codexPeerPath)
-        try legacyPeer.writeToFile(path: legacyPeerPath)
-
-        manager.loadSessions()
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertTrue(visibility.hiddenSessionIDs.isEmpty)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        XCTAssertTrue(finishedPeerPaths.allSatisfy(FileManager.default.fileExists(atPath:)))
-        XCTAssertEqual(
-            manager.cleanupSources.filter { $0.sessionId == codexPeerID }.map(\.projectPath),
-            [codexPeerProjectPath]
-        )
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-
-        try FileManager.default.removeItem(atPath: codexPeerPath)
-        try FileManager.default.removeItem(atPath: legacyPeerPath)
-        manager.loadSessions()
-
-        XCTAssertEqual(visibility.hiddenSessionIDs, [try XCTUnwrap(fastPeer.cctopSessionId)])
-        XCTAssertNil(defaults.object(forKey: ManualSessionVisibilityStore.legacyDefaultsKey))
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertTrue(FileManager.default.fileExists(atPath: fastPeerPath))
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(atPath: historyDir), ["history.json"])
-    }
-
-    @MainActor
-    func testUnresolvedLegacyHideKeepsNewlyArchivedCleanupSourceWhileRecentIsFrozen() throws {
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cctop-unresolved-legacy-new-archive-\(UUID().uuidString)", isDirectory: true)
-        let sessionsURL = root.appendingPathComponent("sessions", isDirectory: true)
-        let historyURL = root.appendingPathComponent("history", isDirectory: true)
-        let finishedProjectURL = URL(fileURLWithPath: #filePath)
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-        try FileManager.default.createDirectory(at: sessionsURL, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: historyURL, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: root) }
-
-        let identityBlockerURL = root.appendingPathComponent("session-identities")
-        try Data("not a directory".utf8).write(to: identityBlockerURL)
-        let hiddenThreadID = "11111111-2222-4333-8444-555555555555"
-        var hidden = codexTerminalSession(sessionId: hiddenThreadID, projectPath: root.path)
-        hidden.cctopSessionId = nil
-        hidden.harnessSessionId = hiddenThreadID
-        let hiddenURL = sessionsURL.appendingPathComponent("codex-\(hiddenThreadID).json")
-        try hidden.writeToFile(path: hiddenURL.path)
-
-        var finished = SessionData(
-            sessionId: "unrelated-finished",
-            projectPath: finishedProjectURL.path,
-            branch: "main",
-            terminal: TerminalInfo(program: "zsh")
-        )
-        finished.cctopSessionId = "22222222-2222-4222-8222-222222222222"
-        finished.source = SessionData.opencodeSource
-        finished.pid = 4_208
-        finished.status = .working
-        let finishedURL = sessionsURL.appendingPathComponent("unrelated-finished.json")
-        try finished.writeToFile(path: finishedURL.path)
-
-        let suiteName = "cctop-unresolved-legacy-new-archive-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["codex:\(hiddenThreadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-
-        var sources = isolatedSessionDataSources(
-            sessionsDir: sessionsURL,
-            manualSessionVisibility: ManualSessionVisibilityStore(defaults: defaults)
-        )
-        sources.codexThreads = StubCodexThreadState(existing: [hiddenThreadID], archived: [])
-        sources.processAlive = { $0.sessionId == hiddenThreadID }
-        let manager = SessionManager(
-            historyManager: HistoryManager(historyDir: historyURL),
-            dataSources: sources,
-            startMonitoring: false
-        )
-
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-        XCTAssertFalse(FileManager.default.fileExists(atPath: finishedURL.path))
-        XCTAssertEqual(try FileManager.default.contentsOfDirectory(at: historyURL, includingPropertiesForKeys: nil).count, 1)
-        XCTAssertEqual(
-            manager.cleanupSources.filter { $0.sessionId == finished.sessionId }.map(\.projectPath),
-            [finishedProjectURL.path]
-        )
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(hiddenThreadID)"]
-        )
-    }
-
-    @MainActor
     func testManualHideImmediatelyRemovesEveryProjectionSharingPermanentIdentity() throws {
         let root = NSTemporaryDirectory() + "cctop-manual-hide-shared-id-\(UUID().uuidString)"
         let sessionsDir = (root as NSString).appendingPathComponent("sessions")
@@ -1502,27 +965,6 @@ final class SessionManagerVisibilityTests: XCTestCase {
         XCTAssertTrue(manager.historyManager.recentProjects.isEmpty)
         XCTAssertTrue(manager.recentResumeTargets.isEmpty)
         XCTAssertTrue(visibility.isHidden(cctopSessionID: hiddenID))
-
-        let threadID = "legacy-hidden-directory-session"
-        let legacySuiteName = "cctop-legacy-hide-recent-directory-\(UUID().uuidString)"
-        let legacyDefaults = try XCTUnwrap(UserDefaults(suiteName: legacySuiteName))
-        defer { legacyDefaults.removePersistentDomain(forName: legacySuiteName) }
-        legacyDefaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let legacyVisibility = ManualSessionVisibilityStore(defaults: legacyDefaults)
-
-        let legacyManager = makeManager(
-            sessionsDir: sessionsPath,
-            historyDir: historyDir,
-            manualSessionVisibility: legacyVisibility
-        )
-
-        XCTAssertTrue(legacyManager.historyManager.recentProjects.isEmpty)
-        XCTAssertTrue(legacyManager.recentResumeTargets.isEmpty)
-        XCTAssertEqual(
-            legacyDefaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-        XCTAssertTrue(legacyVisibility.hiddenSessionIDs.isEmpty)
     }
 
     @MainActor
@@ -3169,75 +2611,7 @@ final class SessionManagerVisibilityTests: XCTestCase {
     }
 
     @MainActor
-    func testGarbageCollectRetainsLegacyCodexManualHideEvidenceBeforeAndAfterMigration() throws {
-        let root = NSTemporaryDirectory() + "cctop-unresolved-legacy-desktop-gc-\(UUID().uuidString)"
-        let sessionsDir = (root as NSString).appendingPathComponent("sessions")
-        let historyDir = (root as NSString).appendingPathComponent("history")
-        try FileManager.default.createDirectory(atPath: sessionsDir, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(atPath: historyDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(atPath: root) }
-
-        let identityBlockerPath = (root as NSString).appendingPathComponent("session-identities")
-        try Data("not a directory".utf8).write(to: URL(fileURLWithPath: identityBlockerPath))
-        let threadID = "66666666-7777-4888-8999-aaaaaaaaaaaa"
-        let cctopSessionID = "66666666-7777-4888-8999-bbbbbbbbbbbb"
-        let old = Date(timeIntervalSinceNow: -SessionManager.lifecycleWindows.retention - 86_400)
-        let sessionPath = (sessionsDir as NSString).appendingPathComponent("\(threadID).json")
-        var session = codexSession(sessionId: threadID, projectPath: "/tmp/unresolved-legacy-desktop")
-        session.cctopSessionId = nil
-        session.harnessSessionId = threadID
-        session.lastActivity = old
-        session.disconnectedAt = old
-        try session.writeToFile(path: sessionPath)
-
-        let suiteName = "cctop-unresolved-legacy-desktop-gc-\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set(["codex:\(threadID)"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
-        let visibility = ManualSessionVisibilityStore(defaults: defaults)
-
-        var sources = isolatedSessionDataSources(sessionsDir: URL(fileURLWithPath: sessionsDir), manualSessionVisibility: visibility)
-        sources.codexThreads = StubCodexThreadState(existing: [threadID], archived: [])
-        sources.desktopAppConnection = DesktopAppConnectionLookup { _ in false }
-        sources.processAlive = { _ in false }
-        let manager = SessionManager(
-            historyManager: HistoryManager(historyDir: URL(fileURLWithPath: historyDir)),
-            dataSources: sources,
-            startMonitoring: false
-        )
-
-        manager.garbageCollectFinished()
-
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        XCTAssertTrue(visibility.hiddenSessionIDs.isEmpty)
-        XCTAssertEqual(
-            defaults.stringArray(forKey: ManualSessionVisibilityStore.legacyDefaultsKey),
-            ["codex:\(threadID)"]
-        )
-
-        session.cctopSessionId = cctopSessionID
-        try session.writeToFile(path: sessionPath)
-        try FileManager.default.removeItem(atPath: identityBlockerPath)
-        manager.loadSessions()
-
-        XCTAssertNil(defaults.object(forKey: ManualSessionVisibilityStore.legacyDefaultsKey))
-        XCTAssertEqual(visibility.hiddenSessionIDs, [cctopSessionID])
-        manager.garbageCollectFinished()
-
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-        manager.loadSessions()
-        XCTAssertEqual(visibility.hiddenSessionIDs, [cctopSessionID])
-        XCTAssertTrue(manager.userSessions.isEmpty)
-        XCTAssertTrue(manager.recentResumeTargets.isEmpty)
-
-        defaults.removeObject(forKey: ManualSessionVisibilityStore.defaultsKey)
-        manager.garbageCollectFinished()
-
-        XCTAssertFalse(FileManager.default.fileExists(atPath: sessionPath))
-    }
-
-    @MainActor
-    func testGarbageCollectRetainsUnreadableLegacyUUIDWhileDurableKeyIsUnresolved() throws {
+    func testGarbageCollectRetainsUnreadableLegacyUUIDWhileManualHideEvidenceExists() throws {
         let root = NSTemporaryDirectory() + "cctop-unreadable-legacy-uuid-gc-\(UUID().uuidString)"
         let sessionsDir = (root as NSString).appendingPathComponent("sessions")
         let historyDir = (root as NSString).appendingPathComponent("history")
@@ -3254,7 +2628,7 @@ final class SessionManagerVisibilityTests: XCTestCase {
         let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
         defer { defaults.removePersistentDomain(forName: suiteName) }
         let cctopSessionID = "77777777-8888-4999-8aaa-cccccccccccc"
-        defaults.set(["codex:unresolved"], forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
+        defaults.set([cctopSessionID], forKey: ManualSessionVisibilityStore.defaultsKey)
         let visibility = ManualSessionVisibilityStore(defaults: defaults)
         let manager = makeManager(
             sessionsDir: sessionsDir,
@@ -3262,12 +2636,6 @@ final class SessionManagerVisibilityTests: XCTestCase {
             manualSessionVisibility: visibility
         )
 
-        manager.garbageCollectFinished()
-
-        XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
-
-        defaults.set([cctopSessionID], forKey: ManualSessionVisibilityStore.defaultsKey)
-        defaults.removeObject(forKey: ManualSessionVisibilityStore.legacyDefaultsKey)
         manager.garbageCollectFinished()
 
         XCTAssertTrue(FileManager.default.fileExists(atPath: sessionPath))
