+++ menubar/CctopMenubarTests/SessionTestSupport.swift
@@ -420,3 +420,63 @@ struct StubClaudeDesktopState: ClaudeDesktopSessionStateProviding {
         snapshot
     }
 }
+
+// MARK: - QA snapshot fixtures
+
+extension SessionData {
+    /// 8 sessions: tests scrolling behavior.
+    static let qaEightSessions: [SessionData] = qaSixSessions + [
+        .mock(id: "7", project: "mobile-app", branch: "release/2.0",
+              status: .waitingPermission, notificationMessage: "Allow Write: /config/prod.json"),
+        .mock(id: "8", project: "analytics", branch: "fix/dashboard", status: .working, lastTool: "Grep", lastToolDetail: "*.ts"),
+    ]
+
+    /// All sessions needing attention (only amber badge visible).
+    static let qaAllAttention: [SessionData] = [
+        .mock(id: "1", project: "web-app", branch: "main", status: .waitingPermission, notificationMessage: "Allow Bash: rm -rf node_modules"),
+        .mock(id: "2", project: "api", branch: "develop", status: .waitingInput, lastPrompt: "Which database migration strategy?"),
+        .mock(id: "3", project: "worker", branch: "main", status: .needsAttention),
+    ]
+
+    /// All sessions idle (only gray badge visible).
+    static let qaAllIdle: [SessionData] = [
+        .mock(id: "1", project: "project-a", branch: "main", status: .idle),
+        .mock(id: "2", project: "project-b", branch: "develop", status: .idle),
+        .mock(id: "3", project: "project-c", branch: "main", status: .idle),
+        .mock(id: "4", project: "project-d", branch: "feature/x", status: .idle),
+    ]
+
+    /// Long project and branch names to test truncation.
+    static let qaLongNames: [SessionData] = [
+        .mock(id: "1", project: "my-very-long-project-name-here",
+              branch: "feature/JIRA-12345-implement-oauth2-refresh-token-rotation",
+              status: .working, lastTool: "Edit",
+              lastToolDetail: "/src/authentication/middleware/refresh-token-handler.ts"),
+        .mock(id: "2", project: "another-extremely-long-name",
+              branch: "fix/bug-that-has-a-really-long-description",
+              status: .waitingInput,
+              lastPrompt: "This is a very long prompt that should be truncated"),
+        .mock(id: "3", project: "short", branch: "m", status: .idle),
+    ]
+
+    /// Long session names to test wrapping (e.g. forked sessions using first message as name).
+    static let qaLongSessionNames: [SessionData] = [
+        .mock(id: "1", project: "cctop",
+              branch: "redesign",
+              sessionName: "Can you use test data to show me what happens if the session name is super long like over 50 characters",
+              status: .working, lastTool: "Edit",
+              lastToolDetail: "/Users/test/projects/cctop/Views/SessionCardView.swift"),
+        .mock(id: "2", project: "cctop",
+              branch: "main",
+              sessionName: "Help me refactor the authentication middleware to support OAuth2 refresh token rotation",
+              status: .idle),
+        .mock(id: "3", project: "blog",
+              branch: "main",
+              status: .idle),
+    ]
+
+    /// Single session.
+    static let qaSingle: [SessionData] = [
+        .mock(id: "1", project: "solo-project", branch: "main", status: .working, lastTool: "Task", lastToolDetail: "Running tests"),
+    ]
+}
